+++ packages/eyes-universal/src/index.ts
@@ -1,1 +1,2 @@
 export * from './universal-server'
+export * from './universal-server-process'