+++ packages/eyes-universal/test/e2e/works.spec.ts
@@ -1,9 +1,9 @@
-import {spawn} from 'child_process'
+import {spawn, fork} from 'child_process'
 
 describe('works', () => {
   const suffixes = {darwin: 'macos', linux: 'linux', win32: 'win'}
 
-  it('works', async () => {
+  it('works with stdout', async () => {
     const server = spawn(`./bin/eyes-universal-${suffixes[process.platform]}`, {
       detached: true,
       shell: process.platform === 'win32' ? 'C:\\Program Files\\Git\\bin\\bash.exe' : '/bin/bash',
@@ -20,6 +20,23 @@ describe('works', () => {
           resolve()
         } else {
           reject(new Error(`Server first line of stdout output expected to be a port, but got "${firstLine}"`))
+        }
+      })
+    }).finally(() => server.kill())
+  })
+
+  it('works with ipc', async () => {
+    const server = fork(`./dist/cli.js`, {detached: true, stdio: 'ignore'})
+    return new Promise<void>((resolve, reject) => {
+      server.on('error', reject)
+
+      const timeout = setTimeout(() => reject(new Error('No output from the server for 10 seconds')), 10000)
+      server.on('message', (data: any) => {
+        clearTimeout(timeout)
+        if (data.name === 'port' && Number.isInteger(data.payload.port)) {
+          resolve()
+        } else {
+          reject(new Error(`Server first message expected to be a port, but got "${JSON.stringify(data)}"`))
         }
       })
     }).finally(() => server.kill())